+++ src/TitleDetails/PiecesActions/index.js
@@ -1,1 +1,2 @@
-export { default } from './ReceivedPiecesActions';
+export { default as ReceivedPiecesActions } from './ReceivedPiecesActions';
+export { default as ExpectedPiecesActions } from './ExpectedPiecesActions';