+++ src/common/resources/index.js
@@ -1,4 +1,4 @@
 export * from './titleResources';
 export * from './orderLineResources';
 export * from './locationResources';
-export * from './pieceResources';
+export * from './receivingResources';