+++ src/common/resources/receivingResources.js
@@ -7,6 +7,8 @@ import {
   ORDER_PIECES_API,
   PIECES_API,
   RECEIVE_API,
+  ITEMS_API,
+  REQUESTS_API,
 } from '../constants';
 
 export const pieceResource = {
@@ -40,4 +42,20 @@ export const receivingResource = {
   clientGeneratePk: false,
   path: RECEIVE_API,
   records: 'toBeReceived',
+};
+
+export const itemsResource = {
+  ...baseManifest,
+  fetch: false,
+  accumulate: true,
+  path: ITEMS_API,
+  records: 'items',
+};
+
+export const requestsResource = {
+  ...baseManifest,
+  fetch: false,
+  accumulate: true,
+  path: REQUESTS_API,
+  records: 'requests',
 };